+++ src/lib/components/ClaudeTimePicker.svelte
