+++ src/lib/components/DoodleIcon.svelte
