+++ src/lib/components/MultiPhotoUploader.svelte
