+++ src/lib/components/NavbarContentLayout.svelte
