+++ src/lib/components/PhotoTaskElement.svelte
