+++ src/lib/components/SimpleDropzone.svelte
