+++ src/lib/components/TextAreaAutoResizing.svelte
