+++ src/routes/+page.svelte
@@ -1,8 +1,8 @@
 <script>
-  import LoginGoogle from '$lib/Reusable/LoginGoogle.svelte'
+  import LoginGoogle from '$lib/components/LoginGoogle.svelte'
   import { hasFetchedUser } from '/src/lib/store'
   import { onMount } from 'svelte'
-  import NavbarContentLayout from '$lib/Reusable/NavbarContentLayout.svelte'
+  import NavbarContentLayout from '$lib/components/NavbarContentLayout.svelte'
 
   let isSoundOff = true
   let VideoElem
