+++ src/routes/[user]/components/Calendar/DayHeader.svelte
@@ -1,7 +1,7 @@
 <script>
-  import CreateTaskDirectly from '$lib/Reusable/CreateTaskDirectly.svelte'
-  import FlexibleDayTask from '$lib/Reusable/FlexibleDayTask.svelte'
-  import DoodleIcon from '$lib/Reusable/DoodleIcon.svelte'
+  import CreateTaskDirectly from '$lib/components/CreateTaskDirectly.svelte'
+  import FlexibleDayTask from '$lib/components/FlexibleDayTask.svelte'
+  import DoodleIcon from '$lib/components/DoodleIcon.svelte'
   import { activeDragItem } from '/src/lib/store'
   import { headerExpanded, isCompact, treesByDate } from './store.js'
   import { DateTime } from 'luxon'
